codatendechat-main/crewai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# CrewAI e dependências
crewai==0.28.0
crewai-tools==0.2.6

# Google Cloud
google-cloud-aiplatform==1.38.1
google-cloud-firestore==2.13.1
firebase-admin==6.2.0

# LangChain
langchain==0.1.0
langchain-google-vertexai==0.1.0
langchain-community==0.0.13

# Utilitários
httpx==0.25.2
